[GOHW-3]: Alpha

+++ cart/internal/domain/item.go
@@ -2,13 +2,13 @@ package domain
 
 type ListItem struct {
 	Sku_id  int64
-	Count   uint16
+	Count   uint64
 	Product Product
 }
 
 type Item struct {
 	Sku_id int64
-	Count  uint16
+	Count  uint64
 }
 
 type ItemsMap map[int64]Item
